Add AOI zoom functionality
